moved resetState file to __tests__ folder
